+++ packages/best_fit/genetic_algorithm.py
@@ -1,6 +1,7 @@
 
 import random
 import numpy as np
+from ..synth_clust import synth_cluster
 import likelihood
 
 #############################################################
@@ -161,9 +162,9 @@ def selection(generation, breed_prob):
     return select_chrom
 
 
-def evaluation(lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
+def evaluation(lkl_method, e_max, err_lst, completeness, max_mag_syn,
                fundam_params, obs_clust, theor_tracks, R_V, ext_coefs,
-               st_dist_mass, N_fc, p_lst, model_done):
+               st_dist_mass, N_fc, cmpl_rnd, err_rnd, p_lst, model_done):
     '''
     Evaluate each model in the objective function to obtain the fitness of
     each one.
@@ -178,11 +179,14 @@ def evaluation(lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
         a_i = fundam_params[1].index(model[1])
         isochrone = theor_tracks[m_i][a_i]
 
+        # Generate synthetic cluster.
+        synth_clust = synth_cluster.main(
+            e_max, err_lst, completeness, max_mag_syn, st_dist_mass, isochrone,
+            R_V, ext_coefs, N_fc, cmpl_rnd, err_rnd, model)
+
         # Call likelihood function for this model.
         # with timeblock(" Likelihood"):
-        lkl = likelihood.main(
-            lkl_method, e_max, bin_mr, err_lst, obs_clust, completeness,
-            max_mag_syn, st_dist_mass, isochrone, R_V, ext_coefs, N_fc, model)
+        lkl = likelihood.main(lkl_method, synth_clust, obs_clust)
 
         # Check if this model was already processed. Without this check here,
         # the extinction/immigration operator becomes useless, since the best
@@ -274,10 +278,10 @@ def num_binary_digits(fundam_params):
     return n_bin, p_delta, p_mins
 
 
-def main(lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
-         fundam_params, obs_clust, theor_tracks, R_V, ext_coefs, st_dist_mass,
-         N_fc, N_pop, N_gen, fit_diff, cross_prob, cross_sel, mut_prob, N_el,
-         N_ei, N_es, flag_print_perc):
+def main(lkl_method, e_max, err_lst, completeness, max_mag_syn, fundam_params,
+         obs_clust, theor_tracks, R_V, ext_coefs, st_dist_mass, N_fc, cmpl_rnd,
+         err_rnd, N_pop, N_gen, fit_diff, cross_prob, cross_sel, mut_prob,
+         N_el, N_ei, N_es, flag_print_perc):
     '''
     Genetic algorithm. Finds the best fit model-observation.
     '''
@@ -305,9 +309,9 @@ def main(lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
 
     # Evaluate initial random solutions in the objective function.
     generation, lkl, model_done = evaluation(
-        lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
-        fundam_params, obs_clust, theor_tracks, R_V, ext_coefs, st_dist_mass,
-        N_fc, p_lst_r, model_done)
+        lkl_method, e_max, err_lst, completeness, max_mag_syn, fundam_params,
+        obs_clust, theor_tracks, R_V, ext_coefs, st_dist_mass, N_fc, cmpl_rnd,
+        err_rnd, p_lst_r, model_done)
 
     # Store best solution for passing along in the 'Elitism' block.
     best_sol = generation[:N_el]
@@ -365,9 +369,9 @@ def main(lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
         # according to the best solutions found.
         # with timeblock("Evaluation"):
         generation, lkl, model_done = evaluation(
-            lkl_method, e_max, bin_mr, err_lst, completeness, max_mag_syn,
+            lkl_method, e_max, err_lst, completeness, max_mag_syn,
             fundam_params, obs_clust, theor_tracks, R_V, ext_coefs,
-            st_dist_mass, N_fc, p_lst_e, model_done)
+            st_dist_mass, N_fc, cmpl_rnd, err_rnd, p_lst_e, model_done)
 
         # *** Extinction/Immigration ***
         # If the best solution has remained unchanged for N_ei