+++ packages/out/synth_cl_file.py
@@ -2,8 +2,8 @@
 from ..synth_clust import synth_cl_plot
 
 
-def main(clp, npd, bf_flag, fundam_params, bin_mr, filters,
-         colors, theor_tracks, R_V, **kwargs):
+def main(clp, npd, bf_flag, fundam_params, filters, colors, theor_tracks,
+         plot_isoch_data, R_V, **kwargs):
     '''
     Create output data file with stars in the best fit synthetic cluster found
     by the 'Best Fit' function.
@@ -11,18 +11,18 @@ def main(clp, npd, bf_flag, fundam_params, bin_mr, filters,
     clp['synth_clst'], clp['shift_isoch'] = [], []
     if bf_flag:
         shift_isoch, synth_clst = synth_cl_plot.main(
-            clp['err_max'], bin_mr, fundam_params, theor_tracks,
+            clp['err_max'], fundam_params, theor_tracks, plot_isoch_data,
             clp['isoch_fit_params'], clp['err_lst'], clp['completeness'],
             clp['max_mag_syn'], clp['st_dist_mass'], R_V, clp['ext_coefs'],
-            clp['N_fc'])
+            clp['N_fc'], clp['cmpl_rnd'], clp['err_rnd'])
 
         # If cluster is not empty.
         if synth_clst:
             # Prepare data.
             mags_cols, e_mags_cols = zip(*synth_clst[0][0]),\
                 zip(*synth_clst[0][1])
-            binar_idx = synth_clst[1]
-            extra_pars = zip(*synth_clst[2])
+            binar_idx = synth_clst[1][0]
+            extra_pars = zip(*synth_clst[1][2:])
             # Prepare header.
             hdr = ['#ID           ']
             hdr += [f[1] + '   ' for f in filters]
@@ -35,7 +35,7 @@ def main(clp, npd, bf_flag, fundam_params, bin_mr, filters,
             with open(npd['synth_file_out'], "w") as f_out:
                 f_out.write(hdr)
                 for i, st in enumerate(mags_cols):
-                    if i in binar_idx:
+                    if binar_idx[i] > 1.:
                         ID = '2' + str(i)
                     else:
                         ID = '1' + str(i)