+++ packages/out/top_tiers.py
@@ -50,8 +50,8 @@ differing\n# >{}% with the best value assigned for said parameter(s).\n\
 
 
 def plot_top_tiers(pd, top_tiers_flo, output_subdir, clust_name, mags,
-                   cols, isoch_fit_params, ext_coefs, N_fc, err_max, err_lst,
-                   completeness, max_mag_syn, st_dist_mass):
+                   cols, isoch_fit_params, ext_coefs, N_fc, cmpl_rnd, err_rnd,
+                   err_max, err_lst, completeness, max_mag_syn, st_dist_mass):
     '''
     Plot all top tiers.
     '''
@@ -67,9 +67,10 @@ def plot_top_tiers(pd, top_tiers_flo, output_subdir, clust_name, mags,
     for mod in top_tiers_flo:
         # Call function to generate synthetic cluster.
         shift_isoch, synth_clst = synth_cl_plot.main(
-            err_max, pd['bin_mr'], pd['fundam_params'], pd['theor_tracks'],
-            isoch_fit_params, err_lst, completeness, max_mag_syn,
-            st_dist_mass, pd['R_V'], ext_coefs, N_fc)
+            err_max, pd['fundam_params'], pd['theor_tracks'],
+            pd['plot_isoch_data'], isoch_fit_params, err_lst, completeness,
+            max_mag_syn, st_dist_mass, pd['R_V'], ext_coefs, N_fc, cmpl_rnd,
+            err_rnd)
 
         # TODO using first magnitude and color defined
         first_col, first_mag = synth_clst[0][0][1], synth_clst[0][0][0]
@@ -93,13 +94,14 @@ def plot_top_tiers(pd, top_tiers_flo, output_subdir, clust_name, mags,
     plt.close("all")
 
 
-def main(npd, cld, pd, err_max, err_lst, ext_coefs, N_fc, completeness,
-         max_mag_syn, st_dist_mass, isoch_fit_params, **kwargs):
+def main(npd, cld, pd, err_max, err_lst, ext_coefs, N_fc, cmpl_rnd, err_rnd,
+         completeness, max_mag_syn, st_dist_mass, isoch_fit_params, **kwargs):
     '''
-    Obtain top tier models, produce data file and output image.
+    Obtain top tier models, produce data file and output image. Don't run
+    if the 'emcee' method was used.
     '''
 
-    if pd['bf_flag']:
+    if pd['bf_flag'] and pd['best_fit_algor'] != 'emcee':
         clust_name, output_subdir = npd['clust_name'], npd['output_subdir']
         all_models = isoch_fit_params[-1]
 
@@ -153,8 +155,8 @@ def main(npd, cld, pd, err_max, err_lst, ext_coefs, N_fc, completeness,
                     plot_top_tiers(
                         pd, top_tiers_flo, output_subdir, clust_name,
                         cld['mags'], cld['cols'], isoch_fit_params, ext_coefs,
-                        N_fc, err_max, err_lst, completeness, max_mag_syn,
-                        st_dist_mass)
+                        N_fc, cmpl_rnd, err_rnd, err_max, err_lst,
+                        completeness, max_mag_syn, st_dist_mass)
                     print("<<Plots for D3 block created>>")
                 except Exception:
                     import traceback